split mesh primitive creation up inside of /primitives
