+++ crates/rmf_site_mesh/src/primitives/mod.rs
@@ -1,0 +1,17 @@
+pub use bevy_render::prelude::*;
+pub use bevy_math::prelude::*;
+
+pub mod halo;
+pub use halo::*;
+
+pub mod pyramid;
+pub use pyramid::*;
+
+pub mod cuboid;
+pub use cuboid::*;
+
+pub mod arrow;
+pub use arrow::*;
+
+pub mod cone;
+pub use cone::*;
